some basic map edits
adds some features and fixes access requirements for restricted parts of the church. Adds aesthetic changes to the church.

Fixes number of keys available for certain areas, fixes the description of keys findable in the game. Adds weapons and drugs to the ganger den and an ability to produce more. Fixes doors to nowhere.